Refactor internal imports and utility usage for consistency
Updated import statements to use new package locations for LoggerFactory, BeanUtils, and BeanRegistrar. Added a private constructor to LocaleUtils to prevent instantiation. Minor code style improvements and cleanup across main and test classes for consistency.
