+++ microsphere-i18n-spring/src/main/java/io/microsphere/i18n/spring/annotation/EnableI18n.java
@@ -24,13 +24,14 @@ import org.springframework.context.annotation.Import;
 import org.springframework.context.support.AbstractApplicationContext;
 
 import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import static io.microsphere.i18n.ServiceMessageSource.COMMON_SOURCE;
+import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
  * Enables the extension for Spring Internationalisation.
@@ -43,8 +44,8 @@ import static io.microsphere.i18n.ServiceMessageSource.COMMON_SOURCE;
  * @see ServiceMessageSourceFactoryBean
  * @since 1.0.0
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+@Retention(RUNTIME)
+@Target({TYPE, ANNOTATION_TYPE})
 @Documented
 @Inherited
 @Import(I18nImportBeanDefinitionRegistrar.class)
