+++ microsphere-i18n-spring/src/main/java/io/microsphere/i18n/spring/util/LocaleUtils.java
@@ -43,4 +43,7 @@ public abstract class LocaleUtils {
         LocaleContext localeContext = getLocaleContext();
         return localeContext == null ? null : localeContext.getLocale();
     }
+
+    private LocaleUtils() {
+    }
 }
